pyproject.toml
[tool.poetry]
name = "graphene-django-filter"
version = "0.2.0"
description = "Advanced filters for Graphene"
authors = ["devind-team <team@devind.ru>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/devind-team/graphene-django-filter"
repository = "https://github.com/devind-team/graphene-django-filter"
keywords = ["django", "graphene", "filter"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Framework :: Django"
]

[tool.poetry.dependencies]
python = ">=3.6.2"
Django = "3.2"
graphene = "2.1.9"
graphene-django = "^2.15.0"
django-filter = "^21.1"
stringcase = "^1.2.0"
anytree = "^2.8.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
flake8-import-order = "^0.18.1"
flake8-docstrings = "^1.6.0"
flake8-builtins = "^1.5.3"
flake8-quotes = "^3.3.1"
flake8-comprehensions = "^3.7.0"
flake8-eradicate = "^1.2.0"
flake8-simplify = "^0.14.5"
pep8-naming = "^0.12.1"
flake8-use-fstring = "^1.3"
flake8-annotations = "^2.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
version_variable = [
    "graphene_django_filter/__init__.py:__version__",
    "pyproject.toml:version"
]
branch = "main"
upload_to_pypi = true
upload_to_release = true
build_command = "pip install poetry && poetry build"
